Домашка № 2.15 выполнена
