+++ index.html
@@ -32,7 +32,7 @@
       <p class="loading-text">Комментарий загружается... </p>
     </div>
   </div>
-</body>
+</body>Identifier 'protectionHtml' has already been declared
 
 
 <script  type="module" src="./script.js"></script>
